requirements.txt
azure-functions
azure-ai-documentintelligence
azure-identity
azure-storage-blob
pandas
openpyxl
jsonschema